+++ rocket.go
@@ -1,6 +1,7 @@
 package rocket
 
 import (
+	"fmt"
 	"log"
 	"net/http"
 	"reflect"
@@ -11,7 +12,7 @@ import (
 
 // Rocket is our service.
 type Rocket struct {
-	port          string
+	port          int
 	router        *router.Route
 	listOfFairing []Fairing
 
@@ -30,8 +31,8 @@ type Rocket struct {
 	onClose func() error
 }
 
-// Ignite initial service by port. The format following native HTTP library `:port_number`
-func Ignite(port string) *Rocket {
+// Ignite initial service by port.
+func Ignite(port int) *Rocket {
 	return &Rocket{
 		port: port,
 		router: router.New(
@@ -90,7 +91,7 @@ func (rk *Rocket) Launch() {
 		f.OnLaunch(rk)
 	}
 	http.HandleFunc("/", rk.ServeHTTP)
-	server := &http.Server{Addr: rk.port, Handler: rk}
+	server := &http.Server{Addr: fmt.Sprintf(":%d", rk.port), Handler: rk}
 	defer func() {
 		if err := server.Close(); err != nil {
 			log.Fatal(err)