+++ rocket_test.go
@@ -12,7 +12,7 @@ func TestOnClose(t *testing.T) {
 		x = 2
 		return nil
 	}
-	err := Ignite("").OnClose(onClose).onClose()
+	err := Ignite(-1).OnClose(onClose).onClose()
 	assert.NoError(t, err)
 	assert.Equal(t, 2, x)
 }