+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Latest
 
 - (#204) `OnClose` takes a function f and runs it after server closed
-- rename `fairingInterface` to `Fairing`, and `Fairing` to `DefaultFairing`
+- (#205) use `int` as **port**, not `string`
+- API rename
+  - `fairingInterface` -> `Fairing`
+  - `Fairing` -> `DefaultFairing`
 - (#198) response: redirect
   ```go
   response.Redirect("/")