+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Latest
 
+## v1.1.0
+
 - (#210) Let guard be a context implements `rocket.Guard`
 - (#204) `OnClose` takes a function f and runs it after server closed
 - (#205) use `int` as **port**, not `string`